+++ src/enamlnative/android/android_date_picker.py
@@ -74,12 +74,12 @@ class AndroidDatePicker(AndroidFrameLayout, ProxyDatePicker):
 
     def set_min_date(self, date):
         #: Convert to long
-        s = long((date - UTC_START).total_seconds() * 1000)
+        s = int((date - UTC_START).total_seconds() * 1000)
         self.widget.setMinDate(s)
 
     def set_max_date(self, date):
         #: Convert to long
-        s = long((date - UTC_START).total_seconds() * 1000)
+        s = int((date - UTC_START).total_seconds() * 1000)
         self.widget.setMaxDate(s)
 
     def set_first_day_of_week(self, day):
